+++ ctrl/config.go
@@ -22,9 +22,10 @@ type Suo5Config struct {
 	RedirectURL      string         `json:"redirect_url"`
 	RawHeader        []string       `json:"raw_header"`
 	DisableHeartbeat bool           `json:"disable_heartbeat"`
+	DisableGzip      bool           `json:"disable_gzip"`
+	TestExit         string         `json:"-"`
 
 	Header                  http.Header                          `json:"-"`
-	TestExit                string                               `json:"-"`
 	OnRemoteConnected       func(e *ConnectedEvent)              `json:"-"`
 	OnNewClientConnection   func(event *ClientConnectionEvent)   `json:"-"`
 	OnClientConnectionClose func(event *ClientConnectCloseEvent) `json:"-"`