+++ ctrl/ctrl.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"crypto/tls"
 	"fmt"
+	"github.com/go-gost/gosocks5"
+	"github.com/go-gost/gosocks5/client"
 	"github.com/go-gost/gosocks5/server"
 	log "github.com/kataras/golog"
 	"github.com/kataras/pio"
@@ -35,6 +37,10 @@ func Run(ctx context.Context, config *Suo5Config) error {
 	err := config.parseHeader()
 	if err != nil {
 		return err
+	}
+	if config.DisableGzip {
+		log.Infof("disable gzip")
+		config.Header.Set("Accept-Encoding", "identity")
 	}
 
 	tr := &http.Transport{
@@ -82,24 +88,27 @@ func Run(ctx context.Context, config *Suo5Config) error {
 
 	log.Infof("header: %s", config.headerString())
 	log.Infof("method: %s", config.Method)
-
-	log.Infof("testing connection with remote server")
-	err = checkMemshell(normalClient, config.Method, config.Target, config.Header.Clone())
+	log.Infof("connecting to target %s", config.Target)
+	result, err := checkConnectMode(config.Method, config.Target, config.Header.Clone())
 	if err != nil {
 		return err
 	}
-	log.Infof("connection to remote server successful")
-	if config.Mode == AutoDuplex || config.Mode == FullDuplex {
-		log.Infof("checking the capability of FullDuplex..")
-		if checkFullDuplex(config.Method, config.Target, config.Header.Clone()) {
-			config.Mode = FullDuplex
+	if config.Mode == AutoDuplex {
+		config.Mode = result
+		if result == FullDuplex {
 			log.Infof("wow, you can run the proxy on FullDuplex mode")
 		} else {
-			config.Mode = HalfDuplex
 			log.Warnf("the target may behind a reverse proxy, fallback to HalfDuplex mode")
 		}
+	} else {
+		if result == FullDuplex && config.Mode == HalfDuplex {
+			log.Infof("the target support full duplex, you can try FullDuplex mode to obtain better performance")
+		} else if result == HalfDuplex && config.Mode == FullDuplex {
+			return fmt.Errorf("the target doesn't support full duplex, you should use HalfDuplex or AutoDuplex mode")
+		}
 	}
-	log.Infof("tunnel created at mode %s!", config.Mode)
+
+	log.Infof("starting tunnel at %s", config.Listen)
 	if config.OnRemoteConnected != nil {
 		config.OnRemoteConnected(&ConnectedEvent{Mode: config.Mode})
 	}
@@ -160,96 +169,120 @@ func Run(ctx context.Context, config *Suo5Config) error {
 			OnClientConnectionClose: config.OnClientConnectionClose,
 		})
 	}()
+	log.Infof("creating a test connection to the remote target")
+	ok := testTunnel(config.Listen, config.Username, config.Password, time.Second*2)
+	time.Sleep(time.Millisecond * 500)
+	if !ok {
+		if !config.DisableGzip {
+			log.Warnf("tunnel test failed")
+			log.Warnf("you can still use the tunnel or retry with --no-gzip to improve compatibility")
+		} else {
+			log.Warnf("tunnel test failed, suo5 can not work on this server, there may be other reverse proxies running on the target")
+		}
+	} else {
+		log.Infof("congratulations! everything works fine")
+	}
 
 	if config.TestExit != "" {
-		time.Sleep(time.Second * 1)
-		if err := testConnection(socks5Addr, config.TestExit, time.Second*15); err != nil {
+		if err := testAndExit(socks5Addr, config.TestExit, time.Second*15); err != nil {
 			return errors.Wrap(err, "test connection failed")
 		}
-		// exit(0)
 		return nil
 	}
+
 	<-ctx.Done()
 	return nil
 }
 
-func checkMemshell(client *http.Client, method string, target string, baseHeader http.Header) error {
-	data := RandString(64)
-	req, err := http.NewRequest(method, target, strings.NewReader(data))
-	if err != nil {
-		return fmt.Errorf("invalid target url, %s", err)
-	}
-	req.Header = baseHeader.Clone()
-	req.Header.Set("Content-Type", ContentTypeChecking)
-	resp, err := client.Do(req)
-	if err != nil {
-		return fmt.Errorf("failed to connect to %s", target)
-	}
-	defer resp.Body.Close()
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return err
-	}
-
-	// 如果是数据末尾的换行，并不影响使用
-	b := strings.TrimRight(string(body), "\r\n")
-	if len(b) < 32 || !strings.HasPrefix(data, b) {
-		header, _ := httputil.DumpResponse(resp, false)
-		log.Errorf("response are as follows:\n%s", string(header)+string(body))
-		return fmt.Errorf("got unexpected body, remote server test failed")
-	}
-
-	return nil
-}
-
-func checkFullDuplex(method string, target string, baseHeader http.Header) bool {
+func checkConnectMode(method string, target string, baseHeader http.Header) (ConnectionType, error) {
 	// 这里的 client 需要定义 timeout，不要用外面没有 timeout 的 rawCient
 	rawClient := rawhttp.NewClient(&rawhttp.Options{
-		Timeout:                3 * time.Second,
+		Timeout:                5 * time.Second,
 		FollowRedirects:        false,
 		MaxRedirects:           0,
 		AutomaticHostHeader:    true,
 		AutomaticContentLength: true,
 		ForceReadAllBody:       false,
 	})
-	data := RandString(64)
+	data := RandString(32)
 	ch := make(chan []byte, 1)
 	ch <- []byte(data)
-	go func() {
-		// timeout
-		time.Sleep(time.Second * 5)
-		close(ch)
-	}()
 	req, err := http.NewRequest(method, target, netrans.NewChannelReader(ch))
 	if err != nil {
-		return false
+		return Undefined, err
 	}
 	req.Header = baseHeader.Clone()
 	req.Header.Set("Content-Type", ContentTypeChecking)
+
+	now := time.Now()
+	go func() {
+		// timeout
+		time.Sleep(time.Second * 3)
+		close(ch)
+	}()
 	resp, err := rawClient.Do(req)
 	if err != nil {
-		return false
+		return Undefined, err
 	}
 	defer resp.Body.Close()
 
-	// 如果此时能立马读取到响应，说明请求没有被缓存, 那么就可以变成全双工的
+	// 如果独到响应的时间在3s内，说明请求没有被缓存, 那么就可以变成全双工的
 	body, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
-		return false
+		return Undefined, err
 	}
+	duration := time.Since(now).Milliseconds()
+
 	b := strings.TrimRight(string(body), "\r\n")
-	return len(b) >= 32 && strings.HasPrefix(data, b)
+	if len(b) >= 32 && strings.HasPrefix(data, b) {
+		if duration < 3000 {
+			return FullDuplex, nil
+		} else {
+			return HalfDuplex, nil
+		}
+	}
+
+	header, _ := httputil.DumpResponse(resp, false)
+	log.Errorf("response are as follows:\n%s", string(header)+string(body))
+	return Undefined, fmt.Errorf("got unexpected body, remote server test failed")
 }
 
-// 检查代理是否真正有效, 只要能按有响应即可，无论目标是否能连通
-func testConnection(socks5 string, remote string, timeout time.Duration) error {
+// 检查代理是否真正有效, 只要能按有响应即可，尝试连一下 server 的 LocalPort, 这里写 0，在 jsp 里有判断
+func testTunnel(socks5, username, password string, timeout time.Duration) bool {
+	addr, _ := gosocks5.NewAddr("127.0.0.1:0")
+	options := []client.DialOption{client.TimeoutDialOption(timeout)}
+	if username != "" && password != "" {
+		options = append(options, client.SelectorDialOption(client.NewClientSelector(url.UserPassword(username, password))))
+	}
+
+	conn, err := client.Dial(socks5, options...)
+	if err != nil {
+		log.Error(err)
+		return false
+	}
+	defer conn.Close()
+	if err := gosocks5.NewRequest(gosocks5.CmdConnect, addr).Write(conn); err != nil {
+		log.Error(err)
+		return false
+	}
+	_ = conn.SetReadDeadline(time.Now().Add(timeout))
+
+	reply, err := gosocks5.ReadReply(conn)
+	if err != nil {
+		log.Error(err)
+		return false
+	}
+	log.Debugf("recv socks5 reply: %d", reply.Rep)
+	return reply.Rep == gosocks5.Succeeded || reply.Rep == gosocks5.ConnRefused
+}
+
+func testAndExit(socks5 string, remote string, timeout time.Duration) error {
 	log.Infof("checking connection to %s using %s", remote, socks5)
 	u, err := url.Parse(socks5)
 	if err != nil {
 		return err
 	}
-	client := http.Client{
+	httpClient := http.Client{
 		Timeout: timeout,
 		Transport: &http.Transport{
 			Proxy: http.ProxyURL(u),
@@ -260,7 +293,7 @@ func testConnection(socks5 string, remote string, timeout time.Duration) error {
 		return err
 	}
 	req.Close = true
-	resp, err := client.Do(req)
+	resp, err := httpClient.Do(req)
 	if err != nil {
 		if os.IsTimeout(err) {
 			return err