+++ gui/frontend/src/Home.vue
@@ -76,7 +76,7 @@
       <span>连接数: {{ status.connection_count }}</span>
       <span>CPU: {{ status.cpu_percent }}</span>
       <span>内存: {{ status.memory_usage }}</span>
-      <span>版本: 0.5.0</span>
+      <span>版本: 0.7.0</span>
     </n-space>
 
     <div class="footer">
@@ -91,14 +91,18 @@
       <n-form label-width="85"
               label-placement="left"
               size="small">
-        <n-grid :cols="2">
-          <n-gi>
+        <n-grid :cols="6">
+          <n-gi span="1">
             <n-form-item label="调试模式">
               <n-checkbox v-model:checked="advancedOptions.debug"></n-checkbox>
             </n-form-item>
           </n-gi>
-
-          <n-gi>
+          <n-gi span="2" offset="1">
+              <n-form-item label-width="100" label="禁用gzip压缩" >
+                  <n-checkbox v-model:checked="advancedOptions.disable_gzip"></n-checkbox>
+              </n-form-item>
+          </n-gi>
+          <n-gi span="1">
             <n-form-item label="禁用心跳包">
               <n-checkbox v-model:checked="advancedOptions.disable_heartbeat"></n-checkbox>
             </n-form-item>
@@ -121,7 +125,7 @@
                    placeholder="用于应对负载均衡，流量将集中转发到这个 url"/>
         </n-form-item>
         <n-form-item label="上游代理">
-          <n-input v-model:value="advancedOptions.upstream_proxy" placeholder="socks5://user:pass@ip:port"/>
+          <n-input v-model:value="advancedOptions.upstream_proxy" placeholder="http(s) or socks5, eg: socks5://user:pass@ip:port"/>
         </n-form-item>
         <n-form-item label="请求头">
           <n-input type="textarea" v-model:value="header"/>
@@ -165,6 +169,7 @@ const formValue = ref<ctrl.Suo5Config>({
   redirect_url: '',
   raw_header: [],
   disable_heartbeat: false,
+  disable_gzip: false,
 })
 
 const advancedOptions = ref<ctrl.Suo5Config>(Object.assign({}, formValue.value))
@@ -192,6 +197,7 @@ const confirmAdvanced = () => {
   formValue.value.raw_header = advancedOptions.value.raw_header
   formValue.value.redirect_url = advancedOptions.value.redirect_url
   formValue.value.disable_heartbeat = advancedOptions.value.disable_heartbeat
+  formValue.value.disable_gzip = advancedOptions.value.disable_gzip
   showAdvanced.value = false
 }
 const formRef = ref<FormInst | null>(null)
@@ -208,6 +214,7 @@ const runAction = async () => {
       return
     }
     checkingAction()
+    console.log(formValue.value)
     await RunSuo5WithConfig(formValue.value)
   } else {
     await Stop()